scheduler: move custom informers for extension
