+++ pkg/scheduler/frameworkext/informer/informers_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package frameworkext
+package informer
 
 import (
 	"context"
